pyproject.toml
[project]
name = "python-boilerplate-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.8",
    "logtail-python>=0.3.2",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-json-logger>=3.2.1",
    "tortoise-orm[asyncpg]>=0.24.0",
]

[dependency-groups]
dev = ["aerich>=0.8.1", "pytest>=8.3.4", "ruff>=0.9.2", "tomlkit>=0.13.2"]

[tool.pytest.ini_options]
pythonpath = "app"

[tool.ruff.lint]
select = ["ANN"]

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
